# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface User {
  id: number;
  email: string;
  role: string;
}

export type UserOrNull = User | null;

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
}

export interface Mod {
  id: number;
  title: string;
  version: string;
  description: string;
  changelog?: string | null;
  imageUrl: string;
  fileUrl: string;
  createdAt: string;
}

export interface CreateModRequest {
  title: string;
  version: string;
  description: string;
  changelog?: string;
  imageUrl: string;
  fileUrl: string;
}

export interface UpdateModRequest {
  title?: string;
  version?: string;
  description?: string;
  changelog?: string;
  imageUrl?: string;
  fileUrl?: string;
}

export interface Fusion {
  id: number;
  name: string;
  type: string;
  recipe: string;
  ability: string;
  imageUrl: string;
  fileUrl?: string | null;
  videoUrl?: string | null;
  createdAt: string;
}

export interface CreateFusionRequest {
  name: string;
  type: string;
  recipe: string;
  ability: string;
  imageUrl: string;
  fileUrl?: string;
  videoUrl?: string;
}

export interface UpdateFusionRequest {
  name?: string;
  type?: string;
  recipe?: string;
  ability?: string;
  imageUrl?: string;
  fileUrl?: string;
  videoUrl?: string;
}

export interface Video {
  id: number;
  title: string;
  thumbnailUrl: string;
  youtubeUrl: string;
  downloadUrl?: string | null;
  description?: string | null;
  createdAt: string;
}

export interface CreateVideoRequest {
  title: string;
  thumbnailUrl: string;
  youtubeUrl: string;
  downloadUrl?: string;
  description?: string;
}

export interface UpdateVideoRequest {
  title?: string;
  thumbnailUrl?: string;
  youtubeUrl?: string;
  downloadUrl?: string;
  description?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  imageUrl?: string | null;
  linkUrl?: string | null;
}

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  imageUrl?: string;
  linkUrl?: string;
}

export interface UpdateCategoryRequest {
  name?: string;
  slug?: string;
  imageUrl?: string;
  linkUrl?: string;
}

export interface CategoryItem {
  id: number;
  categoryId: number;
  title: string;
  description?: string | null;
  imageUrl?: string | null;
  linkUrl?: string | null;
  createdAt: string;
}

export interface CreateCategoryItemRequest {
  title: string;
  description?: string;
  imageUrl?: string;
  linkUrl?: string;
}

export interface UpdateCategoryItemRequest {
  title?: string;
  description?: string;
  imageUrl?: string;
  linkUrl?: string;
}

export interface Setting {
  key: string;
  value: string;
}

export interface SetSettingRequest {
  value: string;
}

export interface UploadResponse {
  url: string;
}

export interface SearchResults {
  mods: Mod[];
  fusions: Fusion[];
  videos: Video[];
}

export type SearchParams = {
  q: string;
};