# lib/api-zod/src/generated/types/createCategoryItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryItemRequest {
  title: string;
  description?: string;
  imageUrl?: string;
  linkUrl?: string;
}