# lib/api-zod/src/generated/types/createCategoryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCategoryRequest {
  name: string;
  slug: string;
  imageUrl?: string;
  linkUrl?: string;
}