# lib/api-zod/src/generated/types/createModRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateModRequest {
  title: string;
  version: string;
  description: string;
  changelog?: string;
  imageUrl: string;
  fileUrl: string;
}