# lib/api-zod/src/generated/types/createVideoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateVideoRequest {
  title: string;
  thumbnailUrl: string;
  youtubeUrl: string;
  downloadUrl?: string;
  description?: string;
}