# lib/api-zod/src/generated/types/fusion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Fusion {
  id: number;
  name: string;
  type: string;
  recipe: string;
  ability: string;
  imageUrl: string;
  fileUrl?: string | null;
  videoUrl?: string | null;
  createdAt: string;
}