# lib/api-zod/src/generated/types/mod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PvZ Fusion Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Mod {
  id: number;
  title: string;
  version: string;
  description: string;
  changelog?: string | null;
  imageUrl: string;
  fileUrl: string;
  createdAt: string;
}